+++ templates/content-single.php
@@ -8,7 +8,7 @@
       <?php the_content(); ?>
     </div>
     <footer>
-      <?php wp_link_pages(array('before' => '<nav class="page-nav"><p>' . __('Pages:', 'forward'), 'after' => '</p></nav>')); ?>
+      <?php wp_link_pages(array('before' => '<nav class="page-nav"><p>' . __('Pages:', LANG_DOMAIN), 'after' => '</p></nav>')); ?>
     </footer>
     <?php comments_template('/templates/comments.php'); ?>
   </article>